+++ vice/src/lib/libresidfp/src/resample/TwoPassSincResampler.h
@@ -39,6 +39,8 @@ namespace reSIDfp
  */
 class TwoPassSincResampler final : public Resampler
 {
+    friend class State;
+
 private:
     std::unique_ptr<SincResampler> const s1;
     std::unique_ptr<SincResampler> const s2;
@@ -72,12 +74,12 @@ public:
             clockFrequency, samplingFrequency, halfFreq, intermediateFrequency);
     }
 
-    bool input(int sample) override
+    bool input(int32_t sample) override
     {
         return s1->input(sample) && s2->input(s1->output());
     }
 
-    int output() const override
+    int32_t output() const override
     {
         return s2->output();
     }